Fix prometheus exporter names
There was a typo in a comment https://github.com/knative/pkg/issues/3224

This change will update the OTel prom exporter metric names to match
what will become the default in the future. If N is the current
SDK version the N+2 is when the new default will apply.

In N+1 a new TranslationStrategy option will be introduced and we
will want to set the strategy to 'UnderscoreEscapingWithSuffixes'.

Then in N+2 we can drop that option as it will become the default
(or leave it explicitly).