Skip disabling keep-alives when shutting down the webhook server (#3305)
My theory is that when this happens the go runtime will close
idle connections. If by chance the API server is making a
request and re-using a connection (that is closing) it will receive
an io.EOF. Thus introducing a flake.

So we now remove this setting change and let all the connections
close during the regular graceful shutdown. We would have
already lame ducked so the webhook shouldn't be receiving new
requests.

Co-authored-by: Dave Protasowski <dprotaso@gmail.com>